Load default organization logo from same h service used for API requests
Ensure that the logo for groups with no organization matches that of groups
which have their organization set to the "default" org by loading the
icon from the same URL that such groups would use
(`{h_url}/organizations/__default__/logo`).

The logo was previously hardcoded in the client, but it turned out that
the hardcoded logo failed to load in Firefox and Chrome because the
content was not URL-escaped. We could just have fixed the encoding, but
this fix ensures that the logo will stay in sync with the h service.